added a module for streaming and added its logic and controller with yt_dlp
