+++ server/src/stream/stream.controller.spec.ts
@@ -1,0 +1,18 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { StreamController } from './stream.controller';
+
+describe('StreamController', () => {
+  let controller: StreamController;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      controllers: [StreamController],
+    }).compile();
+
+    controller = module.get<StreamController>(StreamController);
+  });
+
+  it('should be defined', () => {
+    expect(controller).toBeDefined();
+  });
+});
