+++ server/src/stream/stream.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+import { StreamController } from './stream.controller';
+import { StreamService } from './stream.service';
+
+@Module({
+  controllers: [StreamController],
+  providers: [StreamService]
+})
+export class StreamModule {}
