+++ server/src/stream/stream.service.spec.ts
@@ -1,0 +1,18 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { StreamService } from './stream.service';
+
+describe('StreamService', () => {
+  let service: StreamService;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      providers: [StreamService],
+    }).compile();
+
+    service = module.get<StreamService>(StreamService);
+  });
+
+  it('should be defined', () => {
+    expect(service).toBeDefined();
+  });
+});
